fix: AttemptsLeft should be a byte

+++ internal/api/http.go
@@ -20,7 +20,7 @@ type GuessRequest struct {
 // GameResponse constains the information about a game.
 type GameResponse struct {
 	ID           string  `json:"id"`
-	AttemptsLeft int     `json:"attempts_left"`
+	AttemptsLeft byte    `json:"attempts_left"`
 	Guesses      []Guess `json:"guesses"`
 	WordLength   byte    `json:"word_length"`
 	Solution     string  `json:"solution,omitempty"`
